create new tab collaterals
- move vakuuslaskuri from monitoring to collaterals
